requirements.txt
pandas>=2.1.0
pyarrow>=14.0.0
fastparquet>=2023.10.0
scipy>=1.11.0
scikit-learn>=1.3.0
pyyaml>=6.0.1